Edit del Propietario

+++ MyVet.Web/Controllers/PetTypesController.cs
@@ -125,25 +125,25 @@ namespace MyVet.Web.Controllers
             }
 
             var petType = await _context.PetTypes
-                .FirstOrDefaultAsync(m => m.Id == id);
+                .Include(pt => pt.Pets)
+                .FirstOrDefaultAsync(pt => pt.Id == id);
             if (petType == null)
             {
                 return NotFound();
             }
 
-            return View(petType);
-        }
+            if(petType.Pets.Count > 0)
+            {
+                ModelState.TryAddModelError(string.Empty, "El tipo de mascota no puede eliminarse, tiene mascotas asociadas");
+                return RedirectToAction(nameof(Index));
+            } 
 
-        // POST: PetTypes/Delete/5
-        [HttpPost, ActionName("Delete")]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(int id)
-        {
-            var petType = await _context.PetTypes.FindAsync(id);
             _context.PetTypes.Remove(petType);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
+
+       
 
         private bool PetTypeExists(int id)
         {
